requirement.txt
matplotlib == 3.0.2
numpy == 1.15.4
scipy == 1.2.0
PyYAML == 3.13
elasticsearch == 6.3.1
beautifulsoup4 == 4.7.1